requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
Django==2.1.15
django-celery==3.3.1
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.0
idna==2.10
kombu==3.0.37
protobuf==3.12.4
PyMySQL==0.10.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
